refactor(query): Extract build_view_plan to planner/views.rs
Move view planning logic (206 lines) from builder.rs to views.rs:
- Uses Rust's multiple impl blocks pattern for same-module extraction
- Made lens_store, build_mapping, validate_nested_select_fields pub(super)
- builder.rs reduced from 5,147 to 4,931 lines

This establishes the extraction pattern for further planner splitting.
